+++ src/components/organisms/HomeworksList.stories.js
@@ -1,9 +1,9 @@
 import { storiesOf } from "@storybook/vue";
 import HomeworksList from "@components/organisms/HomeworksList";
-import { homeworks } from "@@/stories/mockData/Homeworks";
+import { homeworks, homeworksTeacher } from "@@/stories/mockData/Homeworks";
 
 storiesOf("0 Vuetify/Homeworks/HomeworksList", module)
-	.add("HomeworksList", () => ({
+	.add("HomeworksList Student", () => ({
 		components: {
 			HomeworksList,
 		},
@@ -11,14 +11,32 @@ storiesOf("0 Vuetify/Homeworks/HomeworksList", module)
 			homeworks,
 		}),
 
-		template: `<v-app><homeworks-list :homeworks="homeworks"/></v-app>`,
+		template: `
+		<v-app>
+			<h1 class="h4">Task Overview for Students</h1>
+			<homeworks-list :homeworks="homeworks" type="student"/>
+		</v-app>`,
+	}))
+	.add("HomeworksList Teacher", () => ({
+		components: {
+			HomeworksList,
+		},
+		data: () => ({
+			homeworks: homeworksTeacher,
+		}),
+
+		template: `
+		<v-app>
+			<h1 class="h4">Task Overview for Teachers</h1>
+			<homeworks-list :homeworks="homeworks" type="teacher"/>
+		</v-app>`,
 	}))
 	.add("HomeworksListLoading", () => ({
 		components: {
 			HomeworksList,
 		},
 		template: `<v-app>
-						<v-skeleton-loader :type="'text'" :max-width="'15%'" />
+						<v-skeleton-loader type="text" :max-width="'15%'" />
 						<v-skeleton-loader
 							v-for="homework of 7"
 							ref="skeleton"